Refactor listing owner check method to utils
