+++ src/controllers/listings.ts
@@ -1,9 +1,10 @@
 import moment from 'moment';
 import { v1 as uuidv1 } from 'uuid';
-import { db } from '../database/db';
+import { db } from '../database';
 import { asyncHandler } from '../middleware';
 import { cleanseData, ErrorResponse, parseSqlUpdateStmt } from '../utils';
 import { isNil } from 'lodash';
+import { checkListingOwner } from '../utils';
 
 /**
  * @desc    Get all listings
@@ -306,8 +307,3 @@ export const searchListings = asyncHandler(async (req, res) => {
         data: rows,
     });
 });
-
-const checkListingOwner = async (userId: string, listingId: string) => {
-    const owner = await db.one<Promise<{ created_by: string }>>('SELECT created_by FROM Listings WHERE listing_id = $1', listingId);
-    return userId === owner.created_by;
-};
