+++ src/utils/listingAuth.ts
@@ -1,0 +1,6 @@
+import { db } from '../database';
+
+export const checkListingOwner = async (userId: string, listingId: string) => {
+    const owner = await db.one<Promise<{ created_by: string }>>('SELECT created_by FROM Listings WHERE listing_id = $1', listingId);
+    return userId === owner.created_by;
+};
